refactor(orders): FE-432 Remove barrel file for `src/components/ui`
