--- apps/storefront/src/components/ui/index.ts
@@ -1,2 +1,0 @@
-export { default as B3Picker } from './B3Picker';
-export { default as B3Select } from './B3Select';
